% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fasta-gff-io.R
\name{findGeneLocusFromToxoDBGFF}
\alias{findGeneLocusFromToxoDBGFF}
\title{findGeneLocusFromToxoDBGFF}
\usage{
findGeneLocusFromToxoDBGFF(geneId, gff)
}
\arguments{
\item{geneId}{toxoDB geneId ex. TGME49_286470}

\item{gff}{path to toxodbgff file}
}
\description{
input toxoDBID return genome locus position (chr, start, end, complement)
~1MB memory for 19MB ToxoDB-28_TgondiiME49.gff 3.10 sec
}

