% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{inputGFF3OutputGenbank}
\alias{inputGFF3OutputGenbank}
\title{inputGFF3OutputGenbank}
\usage{
inputGFF3OutputGenbank(toxoDBGFF)
}
\arguments{
\item{toxoDBGFF}{input file path}
}
\description{
genbank file will be made.
output genbank file is ugene-compatible, may not be compatible for other software.
}
\examples{

}

