% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fasta-gff-io.R
\name{loadAllOnMemotry}
\alias{loadAllOnMemotry}
\title{loadAllOnMemotry}
\usage{
loadAllOnMemotry(geneId, gff)
}
\arguments{
\item{geneId}{toxoDB geneId ex. TGME49_286470}

\item{gff}{path to toxodbgff file}
}
\description{
input toxoDBID return genome locus position (chr, start, end, complement)
around 25MB will be used for 19MB ToxoDB-28_TgondiiME49.gff
}

