% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gff3cleanup.R
\name{toxodb2ugene}
\alias{toxodb2ugene}
\title{toxodb2ugene}
\usage{
toxodb2ugene(toxoDBGFF)
}
\description{
utility for this package functions (not recommended.)
please refer--> https://github.com/Tasu/EpDB2UG/blob/master/build-script/toxodb2ugene.R
}
\examples{
toxodb2ugene(toxoDBGFF)
}

