% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fasta-gff-io.R
\name{modifyGFF4UGENE}
\alias{modifyGFF4UGENE}
\title{modifyGFF4UGENE}
\usage{
modifyGFF4UGENE(toxoDBGFF, UGENEGFF)
}
\arguments{
\item{toxoDBGFF}{input file path}

\item{UGENEGFF}{output file path}
}
\description{
this function edits the base position
}
\examples{
modifyGFF4UGENE(toxoDBGFF, UGENEGFF)
}

