% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fasta-gff-io.R
\name{parsegff3WithSeq}
\alias{parsegff3WithSeq}
\title{parsegff3WithSeq}
\usage{
parsegff3WithSeq(inF, outGFF, outFASTA)
}
\arguments{
\item{inF}{input GFF3 file path}

\item{outGFF}{output GFF file path}

\item{outFASTA}{output fasta file path}
}
\description{
split EupathDB GFF3 with sequence to GFF and FASTA
}
\examples{
inF<-"FILEPATH TO TOXODB GFF3 FILE INCLUDING track data and fasta sequence"
outGFF<-"FILEPATH TO STOR GFF PART"
outFASTA<-"FILEPATH TO STORE FASTA PART"
parsegff3WithSeq(inF,outGFF,outFASTA)
}

