% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gff3cleanup.R
\name{parsegff3WithSeq}
\alias{parsegff3WithSeq}
\title{parsegff3WithSeq}
\usage{
parsegff3WithSeq(inF, outGFF, outFASTA)
}
\arguments{
\item{inF}{outGFF outFASTA}
}
\description{
this function parse gff files
}
\examples{
parsegff3WithSeq(inF,outGFF,outFASTA)
}

