% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{toxodb2ugene}
\alias{toxodb2ugene}
\title{toxodb2ugene}
\usage{
toxodb2ugene(toxoDBGFF)
}
\arguments{
\item{toxoDBGFF}{input file path}
}
\description{
please refer--> https://github.com/Tasu/EpDB2UG/blob/master/build-script/toxodb2ugene.R
output file is ugene-compatible genbank file, may not be compatible for other software.
}
\examples{
toxodb2ugene(toxoDBGFF)
}

