% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{toxodb2ugene}
\alias{toxodb2ugene}
\title{toxodb2ugene}
\usage{
toxodb2ugene(toxoDBGFF)
}
\arguments{
\item{toxoDBGFF}{input file path}
}
\description{
fasta, position fixed gff, genbank file will be made.
output genbank file is ugene-compatible, may not be compatible for other software.
}
\examples{

}

